1. 增加部分json处理器 2. 重构部分代码，整理部分混杂的结构类
