+++ common/src/entity/exception/node.rs
@@ -1,24 +1,4 @@
 #[derive(Debug)]
-pub enum DispatchErr {
-    // 找不到流文件
-    FlowNotFound(String),
-    // 找不到节点
-    NodeNotFound(String),
-    // 运行超限（死循环异常）
-    RunOverLimit,
-    // 补偿流执行失败
-    RedressFailed,
-    // 流执行失败
-    FlowFailed(String),
-    // 流运行需求不满足
-    RequireError(String),
-    // 表达式执行失败
-    EvalExprFailed(String),
-    // 节点执行错误
-    NodeRuntimeError(NodeError),
-}
-
-#[derive(Debug)]
 pub enum NodeError {
     // 找不到第三方的处理器
     ExtNotFound(String),
@@ -70,6 +50,8 @@ pub enum NodeError {
     FileChmodError,
     // 其他文件错误
     FileOtherError(String),
+    // 格式转换错误
+    FormatConvertError(String),
     // 需要管理员权限
     RequirePermission,
     // 脚本执行错误
