+++ common/src/entity/mod.rs
@@ -1,7 +1,5 @@
+pub mod common;
+pub mod extension;
 pub mod flow;
-pub mod net;
-pub mod simx;
-pub mod ext;
-pub mod engine;
-pub mod error;
+pub mod config;
 pub mod exception;
