+++ common/src/exception/flow.rs
@@ -1,4 +1,4 @@
-use crate::entity::error::DispatchErr;
+use crate::entity::exception::dispatch::DispatchErr;
 
 // 流调度错误统一处理器
 pub fn flow_dispatch_err_handler(err: DispatchErr) -> Result<(), DispatchErr> {
