+++ common/src/runtime/thread.rs
@@ -1,4 +1,4 @@
-use crate::entity::simx::SimxThreadSenderStringData;
+use crate::entity::common::SimxThreadSenderStringData;
 use lazy_static::lazy_static;
 use std::collections::HashMap;
 use std::sync::{mpsc, Mutex};
