+++ common/src/thread/extension.rs
@@ -1,5 +1,6 @@
-use crate::entity::ext::Extension;
-use crate::entity::flow::{FlowData, Node};
+use crate::entity::extension::Extension;
+use crate::entity::flow::flow::{FlowData};
+use crate::entity::flow::node::Node;
 
 #[allow(unused_variables)]
 pub fn call(extension: Extension, node: Node, flow_data: &mut FlowData) {}
