+++ engine/src/core/engine/initialization.rs
@@ -4,8 +4,8 @@ use std::path::Path;
 use crate::core::environment::check::env_check;
 use crate::core::flow::interface::load_and_exec_default_flow;
 use crate::core::script::interface::load_and_exec_default_script;
-use engine_common::entity::ext::Extension;
-use engine_common::entity::simx::{SimxFlow, SimxScript};
+use engine_common::entity::extension::Extension;
+use engine_common::entity::common::{SimxFlow, SimxScript};
 use engine_common::logger::interface::{fail, info, success, warn};
 use engine_common::runtime::config::get_simx_config;
 use engine_common::runtime::extension::set_extension_info;
