+++ engine/src/core/environment/interface.rs
@@ -1,5 +1,5 @@
 use crate::core::environment::java::check_java_version;
-use engine_common::entity::flow::Environment;
+use engine_common::entity::flow::flow::Environment;
 
 pub fn check(requirements: Vec<Environment>) -> Result<(), String> {
     for requirement in requirements {
