+++ engine/src/core/flow/resolver/flow.rs
@@ -1,4 +1,4 @@
-use engine_common::entity::flow::Flow;
+use engine_common::entity::flow::flow::Flow;
 use engine_common::logger::interface::fail;
 use std::fs::File;
 use std::io::Read;
