+++ example/flow/init/test.flow
@@ -27,6 +27,9 @@
       "1": {
         "name": "打印信息",
         "handler": "simx.core.debug.print",
+        "common": {
+          
+        },
         "downstream": [
           {
             "expr": "1==1",
