+++ handler/src/handler/core/debug.rs
@@ -1,5 +1,6 @@
-use engine_common::entity::error::NodeError;
-use engine_common::entity::flow::{FlowData, Node};
+use engine_common::entity::exception::node::NodeError;
+use engine_common::entity::flow::flow::{FlowData};
+use engine_common::entity::flow::node::Node;
 use engine_common::logger::interface::{debug, fail, info, warn};
 
 pub fn handle_core_debug(node: Node, flow_data: &mut FlowData) -> Result<(), NodeError> {
