+++ handler/src/handler/files/interface.rs
@@ -1,7 +1,9 @@
 use crate::handler::files::dir::interface::handle_files_dir;
 use crate::handler::files::file::interface::handle_files_file;
-use engine_common::entity::error::NodeError;
-use engine_common::entity::flow::{FlowData, Node};
+use engine_common::entity::exception::node::NodeError;
+use engine_common::entity::flow::flow::{FlowData};
+use engine_common::entity::flow::node::Node;
+use crate::handler::files::json::interface::handle_files_json;
 
 pub fn handle_file(node: Node, flow_data: &mut FlowData) -> Result<(), NodeError> {
     let handler_path: Vec<_> = node.handler.split(".").collect();
@@ -11,7 +13,7 @@ pub fn handle_file(node: Node, flow_data: &mut FlowData) -> Result<(), NodeError
         // 文件相关（创建、重命名、删除、移动）
         "file" => handle_files_file(node, flow_data),
         // Json文件处理
-        "json" => { Ok(()) }
+        "json" => handle_files_json(node, flow_data),
         // Xml文件处理
         "xml" => { Ok(()) }
         _ => {
