+++ handler/src/handler/files/json/mod.rs
@@ -1,0 +1,1 @@
+pub mod interface;
