+++ handler/src/handler/net/http.rs
@@ -1,4 +1,5 @@
-use engine_common::entity::flow::{FlowData, Node};
+use engine_common::entity::flow::flow::{FlowData};
+use engine_common::entity::flow::node::Node;
 use engine_common::logger::interface::warn;
 
 pub async fn handle_net_http(node: Node, flow_data: &mut FlowData) {
