+++ handler/src/handler/net/interface.rs
@@ -1,6 +1,7 @@
 use crate::handler::net::http::handle_net_http;
-use engine_common::entity::error::NodeError;
-use engine_common::entity::flow::{FlowData, Node};
+use engine_common::entity::exception::node::NodeError;
+use engine_common::entity::flow::flow::{FlowData};
+use engine_common::entity::flow::node::Node;
 
 pub async fn handle_net(node: Node, flow_data: &mut FlowData) -> Result<(), NodeError> {
     let handler_path: Vec<_> = node.handler.split(".").collect();
