+++ handler/src/handler/os/shell.rs
@@ -1,5 +1,6 @@
-use engine_common::entity::error::NodeError;
-use engine_common::entity::flow::{FlowData, Node};
+use engine_common::entity::exception::node::NodeError;
+use engine_common::entity::flow::flow::{FlowData};
+use engine_common::entity::flow::node::Node;
 use engine_common::logger::interface::success;
 
 // 控制台/命令行相关（如打印）
